+++ R/gg_text.R
@@ -23,7 +23,7 @@
 #'     y = mpg,
 #'     label = themel,
 #'     y_limits_include = 0,
-#'     y_label = "Miles per gallon",
+#'     y_title = "Miles per gallon",
 #'     col_palette = c(orange, "white", teal),
 #'   )
 #'
@@ -61,7 +61,7 @@ gg_text <- function(
   x_breaks_n = NULL,
   x_expand = NULL,
   x_limits_include = NULL,
-  x_label = NULL,
+  x_title = NULL,
   x_labels = NULL,
   x_position = "bottom",
   x_sec_axis = ggplot2::waiver(),
@@ -71,7 +71,7 @@ gg_text <- function(
   y_breaks_n = NULL,
   y_expand = NULL,
   y_limits_include = NULL,
-  y_label = NULL,
+  y_title = NULL,
   y_labels = NULL,
   y_position = "left",
   y_sec_axis = ggplot2::waiver(),
@@ -81,7 +81,7 @@ gg_text <- function(
   col_breaks_n = 5,
   col_drop = FALSE,
   col_limits_include = NULL,
-  col_label = NULL,
+  col_title = NULL,
   col_labels = NULL,
   col_legend_ncol = NULL,
   col_legend_nrow = NULL,
@@ -103,7 +103,7 @@ gg_text <- function(
   title = NULL,
   subtitle = NULL,
   caption = NULL,
-  label_case = NULL
+  titles_case = NULL
 ) {
   gg_blanket(
     data = data,
@@ -143,7 +143,7 @@ gg_text <- function(
     x_sec_axis = x_sec_axis,
     x_symmetric = x_symmetric,
     x_position = x_position,
-    x_label = x_label,
+    x_title = x_title,
     x_transform = x_transform,
     y_breaks = y_breaks,
     y_expand = y_expand,
@@ -153,7 +153,7 @@ gg_text <- function(
     y_sec_axis = y_sec_axis,
     y_symmetric = y_symmetric,
     y_position = y_position,
-    y_label = y_label,
+    y_title = y_title,
     y_transform = y_transform,
     col_breaks = col_breaks,
     col_drop = col_drop,
@@ -167,7 +167,7 @@ gg_text <- function(
     col_palette_na = col_palette_na,
     col_rescale = col_rescale,
     col_steps = col_steps,
-    col_label = col_label,
+    col_title = col_title,
     col_transform = col_transform,
     facet_axes = facet_axes,
     facet_axis_labels = facet_axis_labels,
@@ -181,7 +181,7 @@ gg_text <- function(
     title = title,
     subtitle = subtitle,
     caption = caption,
-    label_case = label_case,
+    titles_case = titles_case,
     ...
   )
 }