+++ tests/testthat/test-article-go-further.R
@@ -21,7 +21,7 @@ test_that(test_name, {
       y = color,
       width = 0.75,
       x_labels = \(x) x / 1000,
-      x_label = "Count (thousands)",
+      x_title = "Count (thousands)",
       subtitle = "\nDefault order"
     )
 
@@ -37,7 +37,7 @@ test_that(test_name, {
       y = color,
       width = 0.75,
       x_labels = \(x) x / 1000,
-      x_label = "Count (thousands)",
+      x_title = "Count (thousands)",
       subtitle = "\nRe-orderered"
     )
 
@@ -58,7 +58,7 @@ test_that(test_name, {
       y = color,
       width = 0.75,
       x_labels = \(x) x / 1000,
-      x_label = "Count (thousands)",
+      x_title = "Count (thousands)",
       subtitle = "\nUnused levels kept",
     )
 
@@ -71,7 +71,7 @@ test_that(test_name, {
       y = color,
       width = 0.75,
       x_labels = \(x) x / 1000,
-      x_label = "Count (thousands)",
+      x_title = "Count (thousands)",
       subtitle = "\nUnused levels dropped",
     )