+++ tests/testthat/test-positional_scales.R
@@ -18,7 +18,7 @@ test_that(test_name, {
       x_labels = scales::label_currency(),
       x_expand = c(0.5, 0.5),
       x_position = "top",
-      x_label = "Blah",
+      x_title = "Blah",
       caption = ""
     )
 
@@ -38,7 +38,7 @@ test_that(test_name, {
       y_labels = scales::label_currency(),
       y_expand = c(0.1, 0.1),
       y_position = "right",
-      y_label = "Blah",
+      y_title = "Blah",
       caption = ""
     )
 
@@ -58,13 +58,13 @@ test_that(test_name, {
       x_labels = scales::label_currency(),
       x_expand = c(0.5, 0.5),
       x_position = "top",
-      x_label = "Blah",
+      x_title = "Blah",
       y_breaks = scales::breaks_width(1000),
       y_limits_include = 7000,
       y_labels = scales::label_currency(),
       y_expand = c(0.1, 0.1),
       y_position = "right",
-      y_label = "Blah",
+      y_title = "Blah",
       caption = ""
     )